+++ example/example.dart
@@ -3,11 +3,12 @@ import 'package:figma_native_splash/figma_native_splash.dart';
 /// Parse configuration without contacting Figma or modifying an application.
 void main() {
   final config = SplashConfig.parse('''
-figma:
-  phone: https://www.figma.com/design/ExampleFileKey/Launch?node-id=1-2
-platforms: [android, ios]
-ohos:
-  app_splash: false
+splash:
+  figma:
+    phone: https://www.figma.com/design/ExampleFileKey/Launch?node-id=1-2
+  platforms: [android, ios]
+  ohos:
+    app_splash: false
 ''');
   print('Platforms: ${config.platforms.join(', ')}');
   print('Resource prefix: ${config.prefix}');